create new module MeerkatBenchmark
